restructures back-end to accomdate for babel
