+++ .babelrc
@@ -1,5 +1,6 @@
 {
     "presets": [
         "@babel/preset-env"
-    ]
+    ],
+    "plugins": ["@babel/plugin-transform-runtime"]
 }
