+++ .env
@@ -15,7 +15,7 @@ DB_TEST_PORT=3306
 
 # PRODUCTION CREDENTIALS
 DB_PROD_USERNAME=admin
-DB_PROD_PASSWORD=5JKX3kH9Aq40ub66jZ
+DB_PROD_PASSWORD=RHWUbJBNGL0LKN9LiU
 DB_PROD_HOST=aa12qz0ghfocljr.cirtwglmipow.us-east-2.rds.amazonaws.com
 DB_PROD_DATABASE=dewey_prod
 DB_PROD_PORT=3306
